refactor: split index_test.ts into specialized test files
* refactor: decompose monolithic cli/index_test.ts into modular components

- Extract shared test utilities into cli/index_test_base.ts

- Create dedicated test files for help, init, project, compile, and run commands

- Update cli/BUILD to reflect the new modular test structure
