+++ cli/BUILD
@@ -42,10 +42,29 @@ node_modules(
     ],
 )
 
+ts_library(
+    name = "index_test_base",
+    srcs = ["index_test_base.ts"],
+    deps = [
+        "//core",
+        "//protos:ts",
+        "//testing",
+        "@npm//@types/fs-extra",
+        "@npm//@types/js-yaml",
+        "@npm//@types/node",
+        "@npm//fs-extra",
+        "@npm//js-yaml",
+    ],
+)
+
 ts_test_suite(
     name = "tests",
     srcs = [
-        "index_test.ts",
+        "index_help_test.ts",
+        "index_init_test.ts",
+        "index_project_test.ts",
+        "index_compile_test.ts",
+        "index_run_e2e_test.ts",
         "util_test.ts"
     ],
     data = [
@@ -56,17 +75,21 @@ ts_test_suite(
         "@nodejs//:npm",
     ],
     deps = [
+        ":index_test_base",
         "//cli",
+        "//cli/api",
         "//core",
         "//protos:ts",
         "//testing",
         "@npm//@types/chai",
         "@npm//@types/fs-extra",
         "@npm//@types/js-yaml",
+        "@npm//@types/long",
         "@npm//@types/node",
         "@npm//chai",
         "@npm//fs-extra",
         "@npm//js-yaml",
+        "@npm//ts-mockito",
     ],
 )
